# docs/swagger.yaml
basePath: /
definitions:
  models.Project:
    properties:
      id:
        example: "2"
        type: string
      name:
        example: Project2
        minLength: 2
        type: string
      task:
        example: Keyword Spotting
        minLength: 2
        type: string
      taskType:
        example: Audio
        minLength: 2
        type: string
    required:
    - id
    - name
    - task
    - taskType
    type: object
  models.UpdateProjectInput:
    properties:
      name:
        minLength: 2
        type: string
    type: object
info:
  contact: {}
  description: A sample CRUD APIs for managing books.
  title: Project Management API
  version: "1.0"
paths:
  /projects:
    get:
      description: Responds with the list of all projects as JSON.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Project'
            type: array
      summary: Get all projects
      tags:
      - projects
    post:
      consumes:
      - application/json
      description: Takes a JSON input and adds it to the project collection.
      parameters:
      - description: Add project
        in: body
        name: project
        required: true
        schema:
          $ref: '#/definitions/models.Project'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Project'
      summary: Add a new project
      tags:
      - projects
  /projects/{id}:
    delete:
      description: Remove a project from the store by its ID
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "404":
          description: Project not found
      summary: Delete a project
      tags:
      - projects
    get:
      description: Returns a single project based on the ID parameter.
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Project'
        "404":
          description: Not Found
      summary: Get a project by ID
      tags:
      - projects
    put:
      consumes:
      - application/json
      description: Update the details of a project by its ID
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: string
      - description: Updated project object
        in: body
        name: project
        required: true
        schema:
          $ref: '#/definitions/models.UpdateProjectInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Project'
        "400":
          description: Invalid input
        "404":
          description: Project not found
      summary: Update an existing project
      tags:
      - projects
swagger: "2.0"